Move sponsors, add Game Makers, website spacing
Only show City of Helsinki as a main sponsor at the top of the page,
otherwise it's too crowded there. (might try showing them in a
horizontal line later)

Add Game Makers of Finland as a sponsor.

Fix spacing in hero area for smaller screens, make it a bit tighter.